refactor(kolme-runtime): extract runtime outcomes module file (#1962)
